requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.26.0
itsdangerous==2.1.2

# AI Integration
anthropic==0.18.1

# GCP and Infrastructure
google-cloud-storage==2.14.0
google-cloud-resource-manager==1.12.0
google-auth==2.27.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
locust==2.23.1
chaostoolkit==1.17.1
chaostoolkit-kubernetes==0.26.3
chaostoolkit-google-cloud-platform==0.4.1

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
